fix: add ruby linting dependencies (#87)
Because of how super-linter is copying files between their multi-stage docker
images at build time, the apk metadata doesn't exist in the final build
artifact. This causes any apk modification commands such as apk add to purge
the packages which apk is unable to properly manage. In this PR we re-add in
some important packages, and our long term solution will be to move off of
super-linter.

Co-authored-by: JonZeolla <jon.zeolla@seisollc.com>